From Jon's email:
I have a modified version of TArtCalibPPAC.cc. In this version, I have added lines 229-253 and lines 265-285.
These additions take the case that the TSum (for X or Y, respectively) is not in the correct range because of a missing or bad signal.
It checks if ((TX1 or TY1) and TA) or ((TX2 or TY2) and TA) can be used to construct a TSum within the gates inside the configuration file.
Currently, I use a relaxed method to decide if a time is within a good range. For example, to check if TX1 is good, instead of checking:
2*TX1-2*TA >= TXSumMin
I check:
2*TX1-2*TA >= 2*TXSumMin-TXSumMax
and similarly for the maximum limit, I check:
2*TX1-2*TA <=2*TXSumMax-TXSumMin
This is in case TX1 (or other times) are at an extreme of the normal range. In that case, 2*TX1-2*TA may be too small or too large to fit inside the normal range.
In effect, I am tripling the range that I look over to see if TX1 is normal.
If I have good TX1 and TA, I then approximate TX2 as: (TXSumMin+TXSumMax)/2 +2*TA -TX1
This comes from the expression for TSumX= TX1+TX2-2TA, and I take the average of the TSumX constraints as my estimate for TSumX.
There is a possibility that my method will allow some events with multiple particles to pass through this gate.
Clementine used this modified code and it increased the reconstruction efficiency from ~50% to ~90% (more events reconstructed in the expected range) but I have not completed a full study on it.